Fixes: #95
Fixes a crash when a user function or aggregate callable/class gets out of
scope.

The functions/aggregates are now remembered in a dictionary in the Connection
object that gets destroyed on Connection destruction. This way, the refcount
problem is gone, and still the system's global reference count stays constant.
Even if you assign the same function/aggregate multiple times.