Fixes: #130
Deallocate the statement cache before trying to sqlite3_close the database.
This way, the sqlite3_close is successful if the database has not explicitly
been closed before, but the application was relying on the Connection
destructor closing the database.

Otherwise, you would lose file descriptors en masse and run into ugly problems.