Fixes: #197
Under some circumstances it's was not possible to close the connection object.
This happened if you still had cursor objects around with statements that were
no longer in the cached statements pool. This was fixed by finalizing all
statements that were created from the connection instead of only those that are
still found in the connection pool.