Closes #59. Closes #21. Fixes #24. Closes #80.
Completety got rid of statement parsing. We now use SQLite functions to
determine if a statement modifies the database or not. If a statement
modifies the database, then we implicitly start a transaction. This also
means that DDL statements like CREATE TABLE now start transactions.

This is not 100% backwards compatible, but the fix is easy: add commit()
calls to your code after DDL statements. On older pysqlite releases and
when using the sqlite3 module from the standard library, adding the
commit() doesn't hurt.