Allow to use cursors as context managers
This adds methods __enter__ and __exit__ to a cursor, providing it with
context manager functionality. Leaving a cursor context will close the
cursor.

It is important to close cursors in a multi-threaded or multi-process
SQLite situation with writers present, as open cursors will keep read
locks on the database and timely garbage-collection cannot be counted
on.